more updates to jenkinsfile to ensure we only do work on qaw for now

+++ Jenkinsfile
@@ -7,6 +7,7 @@ pipeline {
         registryCredential = 'dockerhub'
         dockerImage = ''
         dockerImageName = ''
+        builderImageName = ''
     }
     agent any
     stages {
@@ -14,12 +15,17 @@ pipeline {
             steps {
                 script {
                     dockerImageName = registry + ":$BRANCH_NAME"
-                    if (env.BRANCH_NAME == 'qaw' || env.BRANCH_NAME == 'prod') {
-                        dockerImage = docker.build(dockerImageName, "-f build/Dockerfile.$BRANCH_NAME .")
+                    builderImageName = registry + ":builder"
+                    if (env.BRANCH_NAME == 'qaw') {
+                        docker.image(builderImageName).pull()
+                        dockerImage = docker.build(dockerImageName, "-f build/Dockerfile.$BRANCH_NAME --cache-from $builderImageName .")
                     }
-                    else {
-                        dockerImage = docker.build(dockerImageName, "-f build/Dockerfile .")
-                    }
+//                     if (env.BRANCH_NAME == 'qaw' || env.BRANCH_NAME == 'prod') {
+//                         dockerImage = docker.build(dockerImageName, "-f build/Dockerfile.$BRANCH_NAME .")
+//                     }
+//                     else {
+//                         dockerImage = docker.build(dockerImageName, "-f build/Dockerfile .")
+//                     }
 
                 }
             }
@@ -27,21 +33,26 @@ pipeline {
         stage('Deploy') {
             steps {
                 script {
-                    docker.withRegistry( '', registryCredential ) {
-                        dockerImage.push()
+                    if (dockerImage) {
+                        docker.withRegistry( '', registryCredential ) {
+                            dockerImage.push()
+                        }
                     }
                 }
             }
         }
         stage('Clean') {
             steps {
-                sh "docker rmi $dockerImageName"
+                if (dockerImage) {
+                    sh "docker rmi $dockerImageName"
+                }
+
             }
         }
         stage('Recreate') {
             steps {
                 script {
-                    if (env.BRANCH_NAME == 'dev') {
+                    if (dockerImage) {
                         httpRequest url: 'http://192.168.0.100:9000/hooks/redeploy', contentType: 'APPLICATION_JSON', httpMode: 'POST', requestBody: """
                             {
                                 "project": {
